+++ src/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 Use inverse order
 
+## v0.11.0 2026-07-31
+* Adding Plugin manager all plugins 
+* Adding custom network builder plugin
+* Adding preprocessing plugin (as normalization)
+* Adding netcdf loader plugin
+
 ## v0.9.2 2026-07-07
 * Added desired output preprocessing + added documentation
 * Fixed the linear_models documentation